The SDL simulator didn't draw the pixels correctly on targets with gray/mono LCD, part 2
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@8625 a1c6a512-1295-4272-9138-f99709370657

+++ uisimulator/sdl/lcd-sdl.c
@@ -129,7 +129,7 @@ void lcd_remote_update_rect(int x_start, int y_start,
         {
             Uint8 *p = (Uint8 *)remote_surface->pixels + y * remote_surface->pitch + x * bpp;
 
-            *(Uint32 *)p = ((lcd_remote_framebuffer[y/8][x] >> (y & 7)) & 1);
+            *(Uint8 *)p = ((lcd_remote_framebuffer[y/8][x] >> (y & 7)) & 1);
         }
 
     SDL_UnlockSurface(remote_surface);
